Small warning fixes.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@32601 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/imagjpeg.cpp
@@ -365,13 +365,13 @@ bool wxJPEGHandler::SaveFile( wxImage *image, wxOutputStream& stream, bool verbo
     if ( image->HasOption(wxIMAGE_OPTION_RESOLUTIONX) &&
          image->HasOption(wxIMAGE_OPTION_RESOLUTIONY) )
     {
-        resX = image->GetOptionInt(wxIMAGE_OPTION_RESOLUTIONX);
-        resY = image->GetOptionInt(wxIMAGE_OPTION_RESOLUTIONY);
+        resX = (UINT16)image->GetOptionInt(wxIMAGE_OPTION_RESOLUTIONX);
+        resY = (UINT16)image->GetOptionInt(wxIMAGE_OPTION_RESOLUTIONY);
     }
     else if ( image->HasOption(wxIMAGE_OPTION_RESOLUTION) )
     {
         resX =
-        resY = image->GetOptionInt(wxIMAGE_OPTION_RESOLUTION);
+        resY = (UINT16)image->GetOptionInt(wxIMAGE_OPTION_RESOLUTION);
     }
     else
     {
